requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-language==1.1.1
googleapis-common-protos==1.6.0
grpcio==1.22.0
idna==2.8
nltk==3.4.4
numpy==1.16.1
pandas==0.25.0
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
